Restructure repo to permit installation
