+++ src/__init__.py
@@ -1,0 +1,5 @@
+"""Define modules"""
+from .emat08_10 import EatonEMAT
+from .dli_dc3 import Dlidc3
+
+__all__ = ["EatonEMAT", "Dlidc3"]
